Decrypt ETSource when deploying

+++ lib/tasks/deploy.rake
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Tasks used during deployment of the application.
 namespace :deploy do
   desc <<-DESC
@@ -7,9 +9,28 @@ namespace :deploy do
   DESC
   task load_etsource: :environment do
     cd(Rails.configuration.etsource_path) do
+      if (current_branch = `git rev-parse --abbrev-ref HEAD`.strip) != 'master'
+        raise 'Cannot load ETSource: requires that the "master" branch be checked out, but ' \
+              "instead found the #{current_branch.inspect} is currently checkout out"
+      end
+
       sh('git fetch origin')
       sh('git reset --hard origin/master')
       sh('git clean -f')
+
+      passphrase = Rails.root.join('config/.etsource_password')
+
+      Pathname.glob('datasets/*/energy_balance.gpg').each do |source|
+        destination = source.dirname.join("#{source.basename('.gpg')}.csv")
+        rm(destination) if destination.file?
+
+        sh(
+          'gpg', '--batch', '--yes', '--ignore-mdc-error',
+          '--passphrase-file', passphrase.expand_path.to_s,
+          '--output', destination.to_s,
+          '--decrypt', source.to_s
+        )
+      end
     end
   end # load_etsource
 end # deploy
